A little POC design re: how we'll actually track processing.
TL;DR - it's the API's job, dummy.

Basically, we'll be using our RDBMS to capture each task to be performed in another table. I doubt this will ever need to scale into the millions, and even if it does, everything I could think of would be time bound. Thus, I doubt the DB will ever be a bottleneck for this sort of application.